# api-spec.yaml
openapi: 3.0.3
info:
  title: Bouldering Navigation API
  version: "1.0"
  description: >
    Source-of-truth API contract for bouldering-navigation.
    Owned by the backend agent — only the backend agent should change this file,
    and only in the same change that implements the corresponding backend code.
    The frontend agent must treat this file as read-only and build against it as-is.
servers:
  - url: http://localhost:8080
paths:
...
  /api/auth/signup:
    post:
      summary: Create a user account and receive a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
        "409":
          description: Duplicate email or nickname
  /api/auth/login:
    post:
      summary: Log in with email and password, receive a bearer token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid email or password
  /api/auth/me:
    get:
      summary: Get the currently authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: Missing or invalid token
...
components:
  schemas:
...
    SignupRequest:
      type: object
      required: [email, password, nickname]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8 }
        nickname: { type: string }
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        tokenType: { type: string, example: Bearer }
        userId: { type: integer, format: int64 }
        nickname: { type: string }
    UserResponse:
      type: object
      properties:
        id: { type: integer, format: int64 }
        email: { type: string }
        nickname: { type: string }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT